Ensure ResourceSynced condition set
There really isn't an "active" or "available" status for Role resources
in IAM. Basically, when the role is successfully created or updated,
it's in sync with whatever the desired state of the role was. So, this
patch adds simple setters after successful create and update code paths
that ensures ResourceSynced condition of True is set on the Role CRs.

Also adds inline policy for managing just roles...

Signed-off-by: Jay Pipes <jaypipes@gmail.com>