add Policy resource (#6)
Adds the Policy resource Kind to the IAM controller. Note that currently
the Update code paths are not supported. This is because the Policy
resource in the IAM API does not support updates. Instead, you need to
call CreatePolicyVersion/DeletePolicyVersion. I will implement this as
custom update code paths in a future PR.

Signed-off-by: Jay Pipes <jaypipes@gmail.com>

aws-controllers-k8s/community#222

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.