Allow Policies to be attached to a Role (#8)
Adds a new `Role.Spec.Policies` field, of type `[]*string` that contains
zero or more Policy ARNs that should be attached to the Role. The
controller uses custom code in a new `pkg/resource/role/hooks.go` file
that evaluates which Policy ARNs to add or remove from a Policy and
calls the IAM ListAttachedRolePolicies, AttachRolePolicy and
DetachRolePolicy API calls.

Signed-off-by: Jay Pipes <jaypipes@gmail.com>

Issue aws-controllers-k8s/community#222

By submitting this pull request, I confirm that my contribution is made under
the terms of the Apache 2.0 license.