release artifacts for v0.0.2 (#9)
Please note the change to the Role CRD is correct. When regenerating
the IAM controller previously, I had to manually update the CRD
definition in `config/crd/bases` but forgot to update the associated
CRD definition in `helm/crds`, which this commit caught.

Signed-off-by: Jay Pipes <jaypipes@gmail.com>

By submitting this pull request, I confirm that my contribution is made under
the terms of the Apache 2.0 license.