add support for updating policy
The IAM Policy resource does not support direct updates. If you want to
modify the policy's Document, you need to call the `CreatePolicyVersion`
API call and set the new policy version as the policy's
DefaultVersionID. Fairly straightforward, as far as non-updateable
resources go.

The trick with IAM Policies is that you cannot have more than 5 policy
versions associated with a user-managed Policy. This means that if you
attempt to make more than 4 calls to `CreatePolicyVersion`, you will get
errors back from IAM. So, instead, you must call `DeletePolicyVersion`
and delete one of the previous policy versions (not the one set as the
default, though).

Similarly, when attmepting to delete a Policy, if there is more than a
single PolicyVersion associated with the Policy, you must first delete
all policy versions *except* the one that is the policy's DefaultVersion
before calling `DeletePolicy`, which is the only way to delete the
policy version that is the default for a policy.

This patch adds support for the above update and delete code paths,
combining all of the various calls to `ListPolicyVersions`,
`GetPolicyVersion` (to get the policy's Document),
`CreatePolicyVersion`, and `DeletePolicyVersion` APIs to make an IAM
Policy resource behave in a "normal Kubernetes way" -- in other words,
the Kubernetes user simply changes the desired state of the Policy
(including the Policy's `Spec.PolicyDocument` field) and the controller
takes care of actuating those changes.

Closes Issue: aws-controllers-k8s/community#1124

Signed-off-by: Jay Pipes <jaypipes@gmail.com>