URL decode `Role` `AssumeRolePolicyDocument` (#28)
Description of changes:
The response from creating or reading a `Role` encodes the `AssumeRolePolicyDocument` with URL encoding (any non-alphanumeric character is converted to hex). When attempting to update the role with the encoded string, the API returns an error. This pull request adds hook code to convert it back into a multiline string so that it matches the original spec.

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.