Fix invalid input exception when updating tags (#30)
Description of changes:
The IAM API will throw an `InvalidInput` exception when attempting to add or delete an empty list of tags. This pull request adds a condition to guard against this.

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.