correct tag comparisons for Role and Policy (#41)
Lists of Tag structs for both the Role and Policy resources were causing
incorrect Delta differences to be detected when the order of the structs
was different. This brings in some code from the sfn-controller project
to properly ignore differences in only ordering for tags.

Issue aws-controllers-k8s/community#1503

Signed-off-by: Jay Pipes <jaypipes@gmail.com>

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license.