Add hooks to set `User` and `Group` paths in their update requests (#46)
The code-generator isn't able to generate correct code for
`newUpdateRequestPayload`. More specifically it doesn't set the
`spec.path` to `input.NewPath`. The reason is that the code-generator
expects the SDK field and CR field names to be similar.

Currently we have a feature that allows the code-generator to set a
field from a different SDK field name. But not the other way around.

We can discuss the above in https://github.com/aws-controllers-k8s/community/issues/1532

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.