Enhance update code path for all the resources
This patch makes changes to the update code path pieces to fix error
handling and minimize the number of API calls made to reconcile a given
resource (tags and policies).

Reading the hooks.go files for `Group` and `Roles` we can observe two
things:
- `syncPolicies` always calls `getPolicies` which is not necessary because
at the beginingg of every reconcilition we already have the "desired" and
"observed" state at hands and we can leverage those to compute the delta
and only make the Add/Remove API Calls.

- Update code for `Group` and `Role` will always call `UpdateGroup` or
`UpdateRole` even if only tags or policies were modified - which not
incorrect but it's an extra API call.