move to fixture-based e2e tests (#55)
I'm going to be adding a number of e2e tests for verifying the resource reference functionality. Figured before I did that, would be good to migrate to pytest fixture-based e2e tests.

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.