Fix Role reconciliation logic for `AssumeRolePolicyDocument`
Addressed a bug in the Role reconciliation mechanism related to
the configuration of a multi-line `AssumeRolePolicyDocument` for
the `Role` resources. The problem was observed as an infinite loop
of requeues and update atempts by the controller. The root cause
was traced back to the behavior of the IAM api, which silently
removed spaces and newline character from the
`AssumeRolePolicyDocument`, leading to faulty comparisons during
the reconciliation process

This commit introduce a solution by removing all instances of
space `(' ')` and newline `('\n')` characters from the targeted
`AssumeRolePolicyDocument`. As aa result, the reconciliation process
now accurately validates and updates the field, eliminating the
previous issue

Use @a-hilaly fork of github.com/micahhausler/aws-iam-policy while waiting for a proper patch/PR

Signed-off-by: Amine Hilaly <hilalyamine@gmail.com>